jour2-docker-mysql/app/requirements.txt
virtualenv
Flask
Redis
mysql-connector-python
sqlalchemy
PyMySQL
pytest